[refactor] 코드 세분화
- src/js 폴더에 js 코드 세분화
