+++ index.html
@@ -6,9 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link rel="stylesheet" href="/src/common/index.css" />
     <link rel="stylesheet" href="/src/styles/main.css" />
-    <script type="module" src="/src/components/currentWeather/currentWeather.js"></script>
-    <script type="module" src="/src/components/playlist/playlist.js"></script>
-    <script type="module" src="/src/components/placeRecommand/woohyeun.js"></script>
+    <script type="module" src="/src/js/main.js"></script>
     <title>날씨</title>
   </head>
   <body>
