+++ src/service/kakaoMap.js
@@ -1,5 +1,7 @@
 import axios from "axios";
 
+const apiKey = import.meta.env.VITE_KAKAO_MAP_API_KEY;
+
 // 현재 위치기반(위도,경도)로 행정구역 명칭 가져오기
 export async function getCurrentLocation(lat, lon) {
   const { data } = await axios.get("https://dapi.kakao.com/v2/local/geo/coord2regioncode.json", {
@@ -8,7 +10,7 @@ export async function getCurrentLocation(lat, lon) {
       x: lon,
     },
     headers: {
-      Authorization: "KakaoAK ", // REST API_KEY
+      Authorization: `KakaoAK ${apiKey}`, // REST API_KEY
     },
   });
