+++ src/service/openWeatherMap.js
@@ -3,11 +3,13 @@ import axios from "axios";
 // axios의 get 메서드를 통해 openWeatherMap API의 날씨데이터를 가져옴
 // 위도,경도를 파라미터로 사용
 
+const apiKey = import.meta.env.VITE_OPENWEATHERMAP_API_KEY;
+
 // 현재날씨 가져오기
 export async function getCurrentWeather(lat, lon) {
   const { data } = await axios.get("https://api.openweathermap.org/data/2.5/weather", {
     params: {
-      appid: "", // API_KEY
+      appid: apiKey, // API_KEY
       lang: "kr",
       units: "metric",
       lat,
@@ -22,7 +24,7 @@ export async function getCurrentWeather(lat, lon) {
 export async function getForecastWeather(lat, lon) {
   const { data } = await axios.get("https://api.openweathermap.org/data/2.5/forecast", {
     params: {
-      appid: "", // API_KEY
+      appid: apiKey, // API_KEY
       lang: "kr",
       units: "metric",
       lat,
