+++ src/styles/main.css
@@ -2,6 +2,7 @@
 @import "/src/components/currentWeather/currentWeather.css";
 @import "/src/components/weeklyforecast/weeklyforecast.css";
 @import "/src/components/playlist/playlist.css";
+@import "/src/components/placeRecommand/woohyeun.css";
 
 body {
   padding: 3.875rem;
